// src/types/Meanings.ts
export type Meaning = {
    id: string
    label: {
        es: string
        en: string
        fr: string
        pt: string
    }
    definitions: {
        es: string
        en: string
        fr: string
        pt: string
    }
    examples: {
        es: string[]
        en: string[]
        fr: string[]
        pt: string[]
    }
}